worker/requirements.txt
prefect
httpx
pydantic
mutagen
pyacoustid
musicbrainzngs
boto3
PyYAML
pytest
